requirements.txt
streamlit==1.28.0
Pillow==10.0.1
openCV-python==4.8.0.76